pyproject.toml
[tool.poetry]
name = "c4v-py"
version = "0.1.0.dev3"
description = "Code for Venezuela python library."
authors = ["Edilmo Palencia <edilmo@gmail.com>"]
license = "Apache-2.0"
readme = "README.md"
homepage = "https://www.codeforvenezuela.org/"
repository = "https://github.com/dieko95/c4v-py"
keywords = [
    "NLP",
    "NLU",
    "Machine Learning",
    "ipython",
    "jupyter",
    "widgets",
    "brat",
    "annotation",
    ]
packages = [
    { include = "c4v", from = "src" },
]
include = ["*.html", "*.ini", "*.json",]

[tool.poetry.dependencies]
python = "^3.8 || ~3.6.2 || ~3.7.0"
pip = "^21.0.0"
tensorflow = "2.5.0"
tensorflow_hub = { extras=["make_image_classifier"], version = "0.8.0" }
tensorflow-probability = "0.10.0"
jupyter = "^1.0.0"
scikit-learn = "0.23.1"
scikit-multilearn = "0.2.0"
pandas = "1.1.1"
nltk = "^3.5"
google-cloud-bigquery = "1.26.1"
Scrapy = "^2.5.0"
beautifulsoup4 = "^4.9.3"
tabulate = "^0.8.9"
click = "7.1.2"
scipy = "1.5.4"
nbconvert = "5.6.1"
traitlets = "4.3.3"
ipykernel = "5.5.5"
ipython = "7.16.1"
zipp = "^3.5.0"
importlib-metadata = "^4.6.1"
importlib-resources = {version = "^5.2.2"}
dataclasses = {version = "^0.8.0", python = ">=3.6.1, <3.7"}
transformers = "^4.9.0"
datasets = "^1.10.2"
torch = "~1.9.0"
dynaconf = "^3.1.4"
mkdocs = "^1.2.2"
transformers-interpret = "^0.5.2"

[tool.poetry.dev-dependencies]
GPUtil = "^1.4.0"
pytest = "^5.4.3"
black = "^19.10b0"
coverage = { extras = ["toml"], version = "^5.1" }
pytest-cov = "^2.10.0"
pytest-mock = "^3.1.1"
zipp = { version = "^3.1.0", python = "~3.6 || ~3.7" }
pylint = "^2.6.0"

[tool.poetry.scripts]
c4v = "c4v.c4v_cli:c4v_cli"

[tool.coverage.paths]
source = ["src", "*/site-packages"]

[tool.coverage.run]
branch = true
source = ["c4v"]

[tool.coverage.report]
show_missing = true
fail_under = 0

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
